# fronted/OPENAPI/ExamSystemApi.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ExamSystemApi",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/v1/students/{page}/{size}": {
      "get": {
        "summary": "findAll",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "分页查询所有学生",
                      "data": {
                        "records": [
                          {
                            "studentId": 0,
                            "studentName": "",
                            "grade": "",
                            "major": "",
                            "clazz": "",
                            "institute": "",
                            "tel": "",
                            "email": "",
                            "pwd": "",
                            "cardId": "",
                            "sex": "",
                            "role": ""
                          }
                        ],
                        "total": 0,
                        "size": 0,
                        "current": 0,
                        "ascs": [
                          ""
                        ],
                        "descs": [
                          ""
                        ],
                        "optimizeCountSql": false,
                        "isSearchCount": false,
                        "searchCount": false,
                        "pages": 0,
                        "asc": [
                          ""
                        ],
                        "desc": [
                          ""
                        ]
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/student/{studentId}": {
      "get": {
        "summary": "findById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "请求成功",
                      "data": {
                        "studentId": 0,
                        "studentName": "",
                        "grade": "",
                        "major": "",
                        "clazz": "",
                        "institute": "",
                        "tel": "",
                        "email": "",
                        "pwd": "",
                        "cardId": "",
                        "sex": "",
                        "role": ""
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "deleteById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "删除成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/studentPWD": {
      "put": {
        "summary": "updatePwd",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "密码更新成功",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/student": {
      "put": {
        "summary": "update",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "更新成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "add",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "添加成功",
                      "data": null
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/api/v1/score/{page}/{size}/{studentId}": {
      "get": {
        "summary": "findById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "studentId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "根据ID查询成绩",
                      "data": {
                        "records": [
                          {
                            "examCode": 0,
                            "studentId": 0,
                            "subject": "",
                            "ptScore": 0,
                            "etScore": 0,
                            "score": 0,
                            "scoreId": 0,
                            "answerDate": ""
                          }
                        ],
                        "total": 0,
                        "size": 0,
                        "current": 0,
                        "ascs": [
                          ""
                        ],
                        "descs": [
                          ""
                        ],
                        "optimizeCountSql": false,
                        "isSearchCount": false,
                        "searchCount": false,
                        "pages": 0,
                        "asc": [
                          ""
                        ],
                        "desc": [
                          ""
                        ]
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/score/{studentId}": {
      "get": {
        "summary": "findById",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "根据ID查询成绩",
                      "data": [
                        {
                          "examCode": 0,
                          "studentId": 0,
                          "subject": "",
                          "ptScore": 0,
                          "etScore": 0,
                          "score": 0,
                          "scoreId": 0,
                          "answerDate": ""
                        }
                      ]
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/score": {
      "post": {
        "summary": "add",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Score",
                "description": ""
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "成绩添加成功",
                      "data": 0
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/api/v1/scores/{examCode}": {
      "get": {
        "summary": "findByExamCode",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "examCode",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResult",
                  "description": "fun.timu.exam.entity.ApiResult"
                },
                "examples": {
                  "1": {
                    "summary": "成功示例",
                    "value": {
                      "code": 200,
                      "message": "查询成功",
                      "data": [
                        {
                          "examCode": 0,
                          "studentId": 0,
                          "subject": "",
                          "ptScore": 0,
                          "etScore": 0,
                          "score": 0,
                          "scoreId": 0,
                          "answerDate": ""
                        }
                      ]
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Student": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "integer",
            "description": ""
          },
          "studentName": {
            "type": "string",
            "description": ""
          },
          "grade": {
            "type": "string",
            "description": ""
          },
          "major": {
            "type": "string",
            "description": ""
          },
          "clazz": {
            "type": "string",
            "description": ""
          },
          "institute": {
            "type": "string",
            "description": ""
          },
          "tel": {
            "type": "string",
            "description": ""
          },
          "email": {
            "type": "string",
            "description": ""
          },
          "pwd": {
            "type": "string",
            "description": ""
          },
          "cardId": {
            "type": "string",
            "description": ""
          },
          "sex": {
            "type": "string",
            "description": ""
          },
          "role": {
            "type": "string",
            "description": ""
          }
        }
      },
...
      "ApiResult": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "错误码，表示一种错误类型\n请求成功，状态码为200"
          },
          "message": {
            "type": "string",
            "description": "对错误代码的详细解释"
          },
          "data": {
            "type": "integer",
            "description": "返回的结果包装在value中，value可以是单个对象"
          }
        }
      },
...
      "Score": {
        "type": "object",
        "properties": {
          "examCode": {
            "type": "integer",
            "description": ""
          },
          "studentId": {
            "type": "integer",
            "description": ""
          },
          "subject": {
            "type": "string",
            "description": ""
          },
          "ptScore": {
            "type": "integer",
            "description": ""
          },
          "etScore": {
            "type": "integer",
            "description": ""
          },
          "score": {
            "type": "integer",
            "description": ""
          },
          "scoreId": {
            "type": "integer",
            "description": ""
          },
          "answerDate": {
            "type": "string",
            "description": ""
          }
        }
      },
...
    },
    "securitySchemes": {}
  },
  "servers": []
}